chore(analytics): update unit tests
