+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/test/java/io/gravitee/apim/infra/domain_service/analytics_engine/permissions/AbstractTest.java
@@ -15,23 +15,10 @@
  */
 package io.gravitee.apim.infra.domain_service.analytics_engine.permissions;
 
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-
-import io.gravitee.rest.api.service.PermissionService;
-import io.gravitee.rest.api.service.v4.ApiAuthorizationService;
 import jakarta.inject.Inject;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.context.SecurityContextImpl;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
@@ -41,30 +28,8 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 @ContextConfiguration(classes = { ResourceContextConfiguration.class })
 @ExtendWith(SpringExtension.class)
-public abstract class AbstractPermissionsTest {
-
-    @Inject
-    PermissionService permissionService;
-
-    @Inject
-    ApiAuthorizationService apiAuthorizationService;
+public abstract class AbstractTest {
 
     @Inject
     ApiAnalyticsQueryFilterDecoratorImpl apiAnalyticsQueryFilterDecorator;
-
-    static final Authentication authentication = mock(Authentication.class);
-
-    @BeforeEach
-    public void setUp() {
-        SecurityContextHolder.setContext(new SecurityContextImpl(authentication));
-    }
-
-    void setAuthenticatedUsername(String username) {
-        when(authentication.getName()).thenReturn(username);
-    }
-
-    void setAuthorities(GrantedAuthority... authorities) {
-        Collection authorityList = new ArrayList<>(Arrays.stream(authorities).toList());
-        when(authentication.getAuthorities()).thenReturn(authorityList);
-    }
 }
